Add files via upload
Client-Server 1: 
This utility includes a separate client and server. 
Client receives a question from the Server "What is your name?".  
The Client sends a reply string to the Server containing a name. 
The Server then sends a reply to he Client, and depending on 
if the name matches the name stored by the Server

Client-Server 2: 
This utility includes a separate client and server. 
The Client socket receives a question from the Server socket  "What is your name?". 
The Client sends a reply string to the Server with the name. 
The Server then sends a reply to the Client depending on if the name matches the 
name stored by the Server
If the name is Joe the Server returns a greeting, otherwise the Client must enter a password
If the password is correct the Client receives a positive response, otherwise it must guess 
the password, until it guesses one of the answers in the array of passwords stored by the Server socket